+++ docs/src/_utils/_templates/class.rst
