+++ docs/src/_utils/_templates/function.rst
