+++ docs/src/_utils/references.bib
